chore: rearrange for readability
